Add support for downloading images
This commits adds support for downloading images via the importd dbus api.

Users can now define an *image_download* block according to which the driver
will try and pull the image before staring the task. In addition it is now also
possible to simply declare the name of an image which is registered with the
machined dbus api in the *image* stanza. When downloading an image which is not
already present on the client, it will be saved with the name configured in
the *image* stanza.