Implement nomad networking
Nomad will set up the NetNS for us, we just have to set it for the
nspawn-container.

This will also mean automatic CNI support in the next nomad release.

I assumed it was safe to remove the IP-table rule generation, as I hope
nomad will make sure that the generated netns is routable correctly.