WIP: Don't lookup machine during startup
The machine might not appear at all (e.g. a failed task) so it's not the
right point to call this.

I noticed only machineName is currently used; with me moving networking
to nomad's responsibility, so there's no need to call DescribeMachine in
the first place, but for getting the main PID which is just
informational