Ensure images can safely be downloaded in parallel
systemd-importd only allows one transfer for each unique URL at a
time. To not run into API errors, we need to ensure we do not try to
download an image from the same URL multiple times at one. We do this
by creating a simple map containing a Mutex for each URL and only
start our download if we can hold the lock for a given URL. This
naively assumes we are the only process making regular use of the
systemd-importd api on the host.

In the future it would probably be better to make use of the built-in
signals in systemd-importd as described here:
https://www.freedesktop.org/wiki/Software/systemd/importd/