Add Makefile
Due to some dependency incompabilities with go 1.15 we need to ensure
binaries are built with go 1.14. To do so and also increase overall
convenience, this commit adds a Makefile to the project.

Calling `make build` will automatically download and install the latest
go 1.14 version and use it to compile the binary.

The Makefile also contains some targets for testing, vendoring etc.