Do not try to recover TaskConfig
When `RecoverTask` is called we initially tried to recover the
`TaskConfig` for a given task. This was blindly copied from the
nomad-driver-skeleton project and it turns out we make no use of it at
all.

Since this also caused Issue #17, we simply get rid of it. Recovering
tasks when a Nomad client is restarted now works again.